requirements.txt
requests>=2.31
beautifulsoup4>=4.12
google-genai>=1.0
fastapi>=0.111
uvicorn>=0.29
psycopg2-binary>=2.9
